requirements.txt
numpy>=1.25
pandas>=2.0           
python-dotenv>=0.21
openai>=1.58.1,<2     
google-genai>=1.15,<2 
httpx>=0.28.1,<1     
transformers>=4.30
torch>=2.0
accelerate>=0.20
sentencepiece>=0.1.97
protobuf>=3.20
pytest>=8,<9
